Rename tracing C++ file.
